+++ pypsa/utils.py
@@ -155,6 +155,7 @@ def pass_none_if_keyerror(func: Callable) -> Callable:
     return wrapper
 
 
+
 def pass_empty_series_if_keyerror(func: Callable) -> Callable:
     @functools.wraps(func)
     def wrapper(*args: Any, **kwargs: Any) -> pd.Series:
@@ -164,3 +165,15 @@ def pass_empty_series_if_keyerror(func: Callable) -> Callable:
             return pd.Series([], dtype=float)
 
     return wrapper
+
+def check_optional_dependency(module_name: str, install_message: str) -> None:
+    """
+    Check if an optional dependency is installed.
+
+    If not, raise an ImportError with an install message.
+    """
+    try:
+        __import__(module_name)
+    except ImportError:
+        raise ImportError(install_message)
+